pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowledge-indexer"
version = "0.1.0"
description = "飞书知识库自动索引工具 - 定时扫描文档、生成摘要标签、构建本地搜索索引、支持自然语言查询"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Knowledge Indexer Team" }]
keywords = ["feishu", "lark", "knowledge-base", "search", "cli", "rag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "openai>=1.30",
    "faiss-cpu>=1.8",
    "numpy>=1.26",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "schedule>=1.2",
]

[project.optional-dependencies]
gpu = ["faiss-gpu>=1.8"]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
knowledge-indexer = "knowledge_indexer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_indexer"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
